auth_api/requirements.txt
alembic==1.11.1
email-validator==2.0.0
fastapi==0.109.1
pydantic-settings
redis==4.5.5
fastapi-limiter==0.1.5
itsdangerous==2.1.2
SQLAlchemy==2.0.17
gunicorn==20.1.0
uvicorn==0.22.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-exporter-jaeger==1.20.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
pydantic==2.6.0
werkzeug==2.2.3
psycopg[binary]==3.1.9
orjson==3.8.5
asyncpg==0.27.0
python-multipart
async-fastapi-jwt-auth[asymmetric]==0.5.1
python-jose==3.3.0
typer~=0.9.0
aioauth-client
asgi-correlation-id==4.3.1
sentry-sdk[fastapi]
